+++ src/index.js
@@ -1,4 +1,4 @@
-import { start } from "./Race.js";
+import { start } from "./domain/Race.js";
 // 입출력 예시
 async function run() {
   await start();
